WIP: First stab at git post-commit hook
The post-commit hook parses the state of the code after each commit to the
`master` branch, builds a tree of method definition "files" using ri-style
naming, and adds a corresponding commit to the orphan `method-log` branch.

* This code assumes the `method-log` branch already exists.

* I've not thought about what happens if the history on `master` is re-written.

* There is no explicit reference between the commits on `master` and those on
the `method-log` branch.

* I'm not sure how tenable it is to use a `Gemfile` and bundler for this hook.
Maybe an in-line `Gemfile` [1] might be better.

* I've duplicated some code from elsewhere in the repo, e.g. `#unindent`.

* I've used `instance_variable_get` on the instance of `MethodFinder`, but it
would be easy to surface this as a public attribute reader method.

[1]: http://bundler.io/whats_new.html#inline